+++ fractran/rust/src/rule.rs
@@ -1,13 +1,13 @@
 // Abstract base class for all transition rules.
 
-use crate::program::{Int, State};
+use crate::program::{BigInt, State};
 
 #[derive(Debug, PartialEq)]
 pub enum ApplyResult {
     // Rule does not apply at all.
     None,
     // Rule applies a finite number of times.
-    Some { num_apps: Int, result: State },
+    Some { num_apps: BigInt, result: State },
     // Rule applies infinitely (proof of non-halting).
     Infinite,
 }
